% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusionMat.R
\name{confusionMatrix}
\alias{confusionMatrix}
\title{Calculate Accuracy Metrics}
\usage{
confusionMatrix(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{A confusion matrix.}
}
\value{
A data frame.
}
\description{
This function takes a confusion matrix as its input and calculates sensitivity, specificity, and kappa scores for classification.
}
