% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectChangers.R
\name{detectChangers}
\alias{detectChangers}
\title{Detect the Accuracy of Change Classification}
\usage{
detectChangers(
  n = 1000,
  t = 3,
  rate = 0.5,
  balance_dir = 0.5,
  balance_res = 0.5,
  strength = 0.5,
  reliable = 1,
  n_samples = 1000
)
}
\arguments{
\item{n}{The number of individuals in the DGP.}

\item{t}{The number of time periods.}

\item{rate}{The percent of individuals changing across the panel.}

\item{balance_dir}{The direction of change, where 0 means all negative change, 1 means all positive change, and values in-between means the percentage of changers changing positively.}

\item{balance_res}{The marginal distribution of the outcome (the percent distribution of 0s and 1).}

\item{strength}{The strength of change in the latent variable.}

\item{reliable}{The reliability score of the outcome measurement.}

\item{n_samples}{The number of simulated datasets.}
}
\value{
A dataframe or a list.
}
\description{
This function builds a simulated dataset and classifies individuals based on their change scores to see whether we can recover the changers from the true DGP.
}
