% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridPlot.R
\name{gridPlot}
\alias{gridPlot}
\title{Plot Grid}
\usage{
gridPlot(
  data,
  method = c("quantile", "normalized"),
  qcut1 = 0.05,
  qcut2 = 0.1,
  qcut3 = 0.2,
  pid_n = NULL,
  ncut1 = 0.05,
  ncut2 = 0.1,
  ncut3 = 0.2
)
}
\arguments{
\item{data}{A data frame generated by \code{gridSearch}}

\item{method}{If \code{method = quantile}, the cuts are based on the quantile of the error distribution. If \code{method = normalized}, cuts are based on the normalized error scores (error / n), which ranges from 0 to 2, for \code{pattern = contingency}, or normalized KS statistics, which ranges from 0 to 1, for \code{pattern = slopes}.}

\item{qcut1}{Cut 1 for the quantile method.}

\item{qcut2}{Cut 2 for the quantile method.}

\item{qcut3}{Cut 3 for the quantile method.}

\item{pid_n}{The number of unique individuals in the original dataframe.}

\item{ncut1}{Cut 1 for the normalized method.}

\item{ncut2}{Cut 2 for the normalized method.}

\item{ncut3}{Cut 3 for the normalized method.}
}
\value{
A \code{ggplot2} object.
}
\description{
This function plots the distribution of error scores resulting from \code{gridSearch}.
}
