% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateChangers.R
\name{simulateChangers}
\alias{simulateChangers}
\title{Simulate Changers and Calculate Accuracy}
\usage{
simulateChangers(
  n = 1000,
  t = 3,
  rate = 0.25,
  balance_dir = 0.5,
  balance_res = 0.5,
  strength = 1,
  reliable = 0.8,
  nrep = 1000,
  seed = 11235,
  workers = 8
)
}
\arguments{
\item{n}{The number of units in the DGP.}

\item{t}{The number of time periods.}

\item{rate}{The percent of units changing across the panel.}

\item{balance_dir}{The direction of change, where 0 codes negative change, 1 codes positive change, and all values in-between codes the percentage of changers changing in the positive direction.}

\item{balance_res}{The marginal distribution of the outcome (effectively regulating the percent distribution of 0s and 1s).}

\item{strength}{The strength of change in the latent variable.}

\item{reliable}{The reliability score of the outcome measurement.}

\item{nrep}{The number of simulation runs.}

\item{seed}{Seed for reproducibility.}

\item{workers}{The number of workers for parallelization (note that parallelization is highly recommended for reasonable duration).}
}
\value{
A data frame.
}
\description{
This function uses parameter values from a given DGP to calculate kappa, sensitivity, and specificity scores in simulated data frames.
}
\section{Examples}{
\strong{Basic \code{\link[=simulateChangers]{simulateChangers()}} call:}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{set.seed(11235)
data <- simulateChangers(n = 100,
                         t = 3,
                         rate = 0.5,
                         balance_dir = 1,
                         balance_res = 0.5,
                         strength = 1,
                         reliable = 0.9,
                         nrep = 100,
                         seed = 11234,
                         workers = 10)
print(data)
#> # A tibble: 100 x 4
#>     sims kappa sensitivity specificity
#>    <int> <dbl>       <dbl>       <dbl>
#>  1     1 0.340       0.432       0.893
#>  2     2 0.119       0.278       0.848
#>  3     3 0.117       0.229       0.885
#>  4     4 0.236       0.362       0.868
#>  5     5 0.370       0.481       0.896
#>  6     6 0.319       0.404       0.906
#>  7     7 0.26        0.38        0.88 
#>  8     8 0.266       0.345       0.952
#>  9     9 0.376       0.545       0.844
#> 10    10 0.120       0.26        0.86 
#> # i 90 more rows
}\if{html}{\out{</div>}}
}

