% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothMatrix.R
\name{smoothMatrix}
\alias{smoothMatrix}
\title{Smooth Matrix
This function takes an \verb{n x m} matrix as its input and averages each cell by using the values from its neighboring cells.}
\usage{
smoothMatrix(mat)
}
\arguments{
\item{mat}{An \verb{n x m} matrix.}
}
\value{
A matrix.
}
\description{
Smooth Matrix
This function takes an \verb{n x m} matrix as its input and averages each cell by using the values from its neighboring cells.
}
