% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varyingSlopes.R
\name{varyingSlopes}
\alias{varyingSlopes}
\title{Calculate Varying Slopes}
\usage{
varyingSlopes(data, pname, tname, yname)
}
\arguments{
\item{data}{A dataframe.}

\item{pname}{The unit identifier.}

\item{tname}{The time identifier.}

\item{yname}{The outcome identifier.}
}
\value{
A data frame.
}
\description{
This function estimates individual-level logistic regression models by regressing binary outcomes on a normalized time variable.
}
