% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDGP.R
\name{buildDGP}
\alias{buildDGP}
\title{Build Panel Data from a DGP}
\usage{
buildDGP(
  n = 1000,
  t = 3,
  rate = 0.5,
  balance_dir = 0.5,
  balance_res = 0.5,
  strength = 0.5,
  reliable = 1,
  export = FALSE,
  status = FALSE
)
}
\arguments{
\item{n}{The number of individuals in the DGP.}

\item{t}{The number of time periods.}

\item{rate}{The percent of individuals changing across the panel.}

\item{balance_dir}{The direction of change, where 0 codes negative change, 1 codes positive change, and all values in-between codes the percentage of changers changing in the positive direction.}

\item{balance_res}{The marginal distribution of the outcome (effectively regulating the percent distribution of 0s and 1s).}

\item{strength}{The strength of change in the latent variable.}

\item{reliable}{The reliability score of the outcome measurement.}

\item{export}{If \code{export = TRUE}, the function exports the simulated dataset.}

\item{status}{if \code{status = TRUE}, the function exports a column for changer status.}
}
\value{
A dataframe or a list.
}
\description{
This function builds a simulated dataset and tabulates the simulated trajectories using the parameters supplied in the function call.
}
