% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDGP.R
\name{buildDGP}
\alias{buildDGP}
\title{Build DGP}
\usage{
buildDGP(
  n = 1000,
  t = 3,
  rate = 0.25,
  balance_dir = 0.5,
  balance_res = 0.5,
  strength = 1,
  reliable = 0.8,
  export = TRUE,
  patterns = FALSE,
  slopes = FALSE
)
}
\arguments{
\item{n}{The number of units in the DGP.}

\item{t}{The number of time periods.}

\item{rate}{The percent of units changing across the panel period.}

\item{balance_dir}{The direction of change, where 0 codes negative change, 1 codes positive change, and all values in-between codes the percentage of changers changing in the positive direction.}

\item{balance_res}{The marginal distribution of the outcome (effectively indexing the 0-1 cut for the latent variable).}

\item{strength}{The strength of change in the latent variable.}

\item{reliable}{The reliability score of the outcome measurement.}

\item{export}{If \code{export = TRUE}, the function exports the simulated dataset.}

\item{patterns}{If \code{patterns = TRUE}, the function exports the concatenated panel patterns.}

\item{slopes}{If \code{slopes = TRUE}, the function exports slope coefficients.}
}
\value{
A list.
}
\description{
This function (1) builds simulated datasets using the parameter inputs in the function call, (2) generates panel patterns by concatenating the outcomes across time periods for each individual, and (3) calculates slope coefficients of change at the individual level.
}
