% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridPlot.R
\name{gridPlot}
\alias{gridPlot}
\title{Plot DGP Grids}
\usage{
gridPlot(data, cut1 = 0.05, cut2 = 0.1, cut3 = 0.2)
}
\arguments{
\item{data}{A data frame generated by \code{gridSearch}.}

\item{cut1}{Cut error values for DGPs: Group 1.}

\item{cut2}{Cut error values for DGPs: Group 2.}

\item{cut3}{Cut error values for DGPs: Group 3.}
}
\value{
A \code{ggplot2} object.
}
\description{
This function plots the distribution of error scores resulting from \code{gridSearch}.
}
