% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridPlot.R
\name{gridPlot}
\alias{gridPlot}
\title{Plot DGP Parameters}
\usage{
gridPlot(data, type = c("posterior", "conditional"))
}
\arguments{
\item{data}{A data frame generated by \code{gridSearch}.}

\item{type}{If \code{posterior}, plots a density plot for each non-fixed parameter. If \code{conditional}, plots the accepted samples by different balance levels.}
}
\value{
A \code{ggplot2} object.
}
\description{
This function plots the distribution of accepted samples.
}
