% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearch.R
\name{gridSearch}
\alias{gridSearch}
\title{Grid Search}
\usage{
gridSearch(
  data,
  yname,
  tname,
  pname,
  pattern = c("contingency", "slopes"),
  step1 = 30,
  step2 = 120,
  step3 = 480,
  reliability = 0.9,
  seed = 11235
)
}
\arguments{
\item{data}{A dataframe.}

\item{yname}{The outcome identifier.}

\item{tname}{The time identifier.}

\item{pname}{The unit identifier.}

\item{pattern}{If \code{pattern = contingency}, the calculations are based on the use of a contingency table of panel patterns; if \code{pattern = slopes}, the calculations are based on individual slope coefficients across time.}

\item{step1}{The number of simulated datasets in Step 1.}

\item{step2}{The number of simulated datasets in Step 2.}

\item{step3}{The number of simulated datasets in Step 3.}

\item{reliability}{The reliability score of the outcome measurement.}

\item{seed}{Set seed.}
}
\value{
A data frame.
}
\description{
This function uses a three-step grid search algorithm to calculate the extent to which a dataframe can be approximated by a known DGP.
}
