% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateChangers.R
\name{simulateChangers}
\alias{simulateChangers}
\title{Simulate Changers}
\usage{
simulateChangers(
  n = 1000,
  t = 3,
  rate = 0.25,
  balance_dir = 0.5,
  balance_res = 0.5,
  strength = 1,
  reliable = 0.8,
  nrep = 1000,
  seed = 11235,
  workers = 8
)
}
\arguments{
\item{n}{The number of units in the DGP.}

\item{t}{The number of time periods.}

\item{rate}{The percent of units changing across the panel period.}

\item{balance_dir}{The direction of change, where 0 codes negative change, 1 codes positive change, and all values in-between codes the percentage of changers changing in the positive direction.}

\item{balance_res}{The marginal distribution of the outcome (effectively indexing the 0-1 cut for the latent variable).}

\item{strength}{The strength of change in the latent variable.}

\item{reliable}{The reliability score of the outcome measurement.}

\item{nrep}{The number of simulation runs.}

\item{seed}{Set seed.}

\item{workers}{The number of workers for parallelization.}
}
\value{
A data frame.
}
\description{
This function gets several DGP parameters and calculates kappa, sensitivity, and specificity scores for classifying individual changers in simulated data frames.
}
